Change logic to enable/disable Time runnable.
Previously, when a view was detached, we removed the view from
the callback list and if this list was empty, we stopped the runnable.
In fact, a timer can run even if there is no view associated to its.
The logic was changed to check 'isActivated' parameters of TimerStates
and check if there is at list one activated to star the runnable or
zero to stop it.