[TimeManager] Refactor to add startTimer and stopTimer
Previously, we needed to attach a callback to TimeManager to start
"update mechanism". To avoid this bad behavior, we add a startTimer
and stopTimer methods with a TimerState as parameter. To be notified
of an update, register methods are still here.

Now views always need to be registered to TimeManager in case we start
a Timer from anywhere.