[CreateTimer] Fix bug relative to directly start timer
Previously, the ID of the timer was added in a dictionnary has a key
to reference last updated time for an active timer.

When we want to directly start a timer, its ID was added in the dictionnary but
its value was 0 because we have a valid ID only after the timer creation in the
database (due to asynchronous behavior). When we loop on the list of activated timer,
timer ID wasn't found and the app crashed.

Now the key is the Timer itself and the copy method has been overrided to remove the check
on the reference. We only check Timer variables now.