[Project] Distinguish version code of installed app from instant app
Both application can't have the same version code. For now, by default, instant app increment the version code of the installed app but there is no check if the install app increment was of just 1 or 2.

It will fail if it's just 1 and if an instant app has been published because it will used the same version code.